pyproject.toml
[project]
name = "viralQC"
version = "1.0.0"
description = "A tool and package for quality control of consensus virus genomes."
authors = [
    { name = "Filipe Zimmer Dezordi", email = "zimmer.filipe@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.8,<3.12"
license = { text = "MIT" }
dependencies = [
    "typer>=0.16",
    "colorlog>=6.3",
    "pyyaml>=6.0",
    "pandas>=2.2",
    "python-multipart==0.0.20",
    "biopython==1.86"
]

keywords = ["virus", "quality-control", "python", "snakemake"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Intended Audience :: Science/Research",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Utilities",
    "Development Status :: 3 - Alpha"
]

[project.optional-dependencies]
cli = [
    "typer>=0.16",
    "colorlog>=6.3"
]
dev = [
    "black>=24.3",
    "pytest>=8.0",
]

[project.scripts]
vqc = "viralqc.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.setuptools.package-data]
viralqc = ["config/datasets.yml"]

[project.urls]
Homepage = "https://github.com/InstitutoTodosPelaSaude/viralQC"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
